+++ doc/Changelog.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 * `Pipeline::expected_inputs` and `Pipeline::expected_outputs` now take the parameters as arguments (in case the expected tensors depend on them) and return iterators for greater flexibility.
 
+* Add `inspect` binary that can be used to display various model properties from metadata and input/output tensors
+
 ### Fixed
 
 * Fixed error message in case of expected/actual tensors mismatch.